MYR-206 Add neighborhood fallback to geocoder PlaceName ranking
Broadens the Mapbox request types to poi,address,neighborhood and adds
a tier-ranked PlaceName selection: a POI within the 50m threshold wins;
otherwise the containing neighborhood's name is used so residential
drive endpoints (the common case) get a human-readable label; otherwise
PlaceName stays empty and clients fall back to the address.

Tier ranking is order-independent: candidates are collected per tier in
one pass and resolved afterwards, so a neighborhood feature listed
before a qualifying POI can never shadow the POI. Within a tier the
first feature in Mapbox's relevance order wins (unchanged POI rule).

City-level types are the hard floor: locality/place features are never
requested, never used, and a dual-tagged neighborhood+locality feature
is excluded, so a bare city name can never leak into PlaceName (an
intra-city drive would render "Dallas -> Dallas", rejected by client
QA). Address extraction is unchanged.

Ranking logic moves to ranking.go — the fallback pushed geocode.go past
the 300-line file limit.
