+++ internal/geocode/geocode.go
@@ -55,7 +55,16 @@ const poiDistanceThreshold = 50.0
 
 // Result holds the output of a reverse geocode lookup.
 type Result struct {
-	// PlaceName is the short place name (e.g. "Thompson Hotel").
+	// PlaceName is the short place name. A POI within
+	// poiDistanceThreshold wins (e.g. "Thompson Hotel"); otherwise the
+	// containing neighborhood's name is used (e.g. "Preston Highlands")
+	// so residential drive endpoints still get a human-readable label.
+	// Empty when neither is available — consumers fall back to Address.
+	//
+	// PlaceName is NEVER a bare city/locality name: an intra-city drive
+	// would render "Dallas → Dallas", which client QA explicitly
+	// rejected (MYR-206). Neighborhood is the floor of the fallback
+	// chain.
 	PlaceName string
 	// Address is the full street address
 	// (e.g. "506 San Jacinto Blvd, Austin, TX 78701").
@@ -142,8 +151,17 @@ func (g *MapboxGeocoder) ReverseGeocode(ctx context.Context, lat, lng float64) (
 	// limit=5 gives us a small candidate set so we can prefer a nearby
 	// POI over the closest address-only feature when one is in walking
 	// range; see the post-decode ranking below.
+	//
+	// types includes "neighborhood" (MYR-206) so buildResult can fall
+	// back to the containing neighborhood's name when no POI is within
+	// walking range — the common residential-endpoint case. "locality"
+	// (and coarser city-level types) are deliberately NOT requested:
+	// buildResult would only ever discard them (a bare city name in
+	// PlaceName renders "Dallas → Dallas" on intra-city drives, rejected
+	// by client QA), so requesting them wastes candidate slots in the
+	// limit=5 window.
 	url := fmt.Sprintf(
-		"https://api.mapbox.com/geocoding/v5/mapbox.places/%f,%f.json?access_token=%s&limit=5&types=poi,address",
+		"https://api.mapbox.com/geocoding/v5/mapbox.places/%f,%f.json?access_token=%s&limit=5&types=poi,address,neighborhood",
 		lng, lat, g.token,
 	)
 
@@ -173,67 +191,6 @@ func (g *MapboxGeocoder) ReverseGeocode(ctx context.Context, lat, lng float64) (
 	}
 
 	return buildResult(lat, lng, data.Features), nil
-}
-
-// buildResult selects the best PlaceName and Address from a Mapbox
-// response.
-//
-// PlaceName: returns the FIRST POI feature (in Mapbox's relevance order)
-// whose center is within poiDistanceThreshold of the queried coord. If
-// no POI qualifies, PlaceName is empty so downstream consumers fall
-// back to the street address — Mapbox's "address" features carry only
-// the street name in Text (e.g. "Tributary Way"), which makes a poor
-// label for a residential drive.
-//
-// Address: prefers the first address feature's full place_name; if none
-// is present, falls back to features[0].place_name so we never return
-// an empty address when the API gave us something.
-func buildResult(lat, lng float64, features []mapboxFeature) *Result {
-	var (
-		placeName   string
-		addressLine string
-		haveAddress bool
-	)
-
-	for _, f := range features {
-		if placeName == "" && isPOI(f.PlaceType) {
-			featLng, featLat := f.Center[0], f.Center[1]
-			if haversineMeters(lat, lng, featLat, featLng) <= poiDistanceThreshold {
-				placeName = f.Text
-			}
-		}
-		if !haveAddress && containsType(f.PlaceType, "address") {
-			addressLine = f.PlaceName
-			haveAddress = true
-		}
-	}
-
-	if !haveAddress {
-		addressLine = features[0].PlaceName
-	}
-
-	return &Result{
-		PlaceName: placeName,
-		Address:   addressLine,
-	}
-}
-
-// isPOI reports whether a Mapbox feature's place_type list includes the
-// "poi" tag.
-func isPOI(types []string) bool {
-	return containsType(types, "poi")
-}
-
-// containsType reports whether want appears in the given place_type
-// list. Mapbox features routinely carry multiple types
-// (e.g. ["poi", "landmark"]), so we cannot rely on the first element.
-func containsType(types []string, want string) bool {
-	for _, t := range types {
-		if t == want {
-			return true
-		}
-	}
-	return false
 }
 
 // NoopGeocoder always returns ErrNoResult, effectively disabling
